Use tiny_sort instead of slice::sort_unstable
This saves about 2% binary size on Linux for a size optimized hello world
program. The standard library unstable sort is optimized to handle large inputs
efficiently and adapt to various patterns. However the amount of symbols that
are sorted are not that large and the gain in binary-size reduction seems worth
it.